update role fixed

+++ db/query.js
@@ -15,7 +15,6 @@ class DB {
         return this.connection.promise().query('SELECT role.id AS role_id, role.title, department.name AS department, salary FROM role LEFT JOIN department ON department.id=role.department_id')
         };
     //query for all employees==all data ids, first last title, departments, salaries, managers
-    //!! NEED id NEED MANAGER NAMES
     viewEmployees() {
         return this.connection.promise().query('SELECT employee.id, employee.first_name, employee.last_name, role.title, department.name AS department, role.salary, CONCAT(manager.first_name, " ", manager.last_name) AS manager FROM employee LEFT JOIN role ON employee.role_id=role.id LEFT JOIN department ON role.department_id=department.id LEFT JOIN employee AS manager ON manager.id=employee.manager_id');
     };
@@ -25,29 +24,27 @@ class DB {
     }
     deleteDepartment(department) {
         return this.connection.promise().query('DELETE FROM department WHERE id =? ', department);
-
     }
     //query add role==asked to add role, then asked name, salary, department for role, added to database
     makeRole(role) {
         return this.connection.promise().query('INSERT INTO role SET ?', role);
     }
-    deleteRole() {
-        return this.connection.promise().query('DELETE FROM ')
+    deleteRole(title) {
+        return this.connection.promise().query('DELETE FROM role WHERE id = ?', title);
     }
     //query to add employee= first last role and manager and that employee is added to database
     addNewEmployee(employee) {
         return this.connection.promise().query('INSERT INTO employee SET ?', employee);
     }
-    deleteEmployee() {
-        return this.connection.prmoise().query('')
+    findDepartmentSalary(roleId) {
+        return this.connection.promise().query('SELECT role.salary, department.name AS department FROM role LEFT JOIN department ON department_id=department.id WHERE role.id=?', roleId);
     }
-    //value obj in inquirer, name goes to user, value is your return
+    deleteEmployee(name) {
+        return this.connection.promise().query('DELETE FROM employee WHERE id = ?', name);
+    }
     //query to update an employee role=employee to update and their new role and updated to database
-    //!!!! FIRST NAME AND LAST NAME ====NAME
-    updateEmployeeRole(update, updateRole) {
-        return this.connection.promise().query('UPDATE employee SET role_id = ? WHERE id = ?', [update, updateRole])
-        // SELECT id AS id, CONCAT(first_name, + ' ', last_name) AS name
-    // FROM employee; gives id and Employee name
+    updateEmployeeRole(updateRole, update) {
+        return this.connection.promise().query('UPDATE employee SET role_id = ? WHERE employee.id = ?', [updateRole, update])
     }
 }
